move exec to installer
