+++ pkg/fleet/internal/bootstrap/bootstrap_windows.go
@@ -11,15 +11,16 @@ package bootstrap
 import (
 	"context"
 	"fmt"
-	"github.com/DataDog/datadog-agent/pkg/fleet/internal/msi"
 	"os"
 	"path/filepath"
+
+	"github.com/DataDog/datadog-agent/pkg/fleet/internal/msi"
 
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/paths"
 
+	iexec "github.com/DataDog/datadog-agent/pkg/fleet/installer/exec"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/oci"
-	iexec "github.com/DataDog/datadog-agent/pkg/fleet/internal/exec"
 )
 
 func install(ctx context.Context, env *env.Env, url string, experiment bool) error {
